requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (SQLite for MVP, upgrade to PostgreSQL later)
sqlalchemy==2.0.23

# Geospatial
rasterio==1.3.9
Pillow==10.1.0

# Cloud Services
cloudinary==1.36.0
google-generativeai==0.3.1

# Utilities
httpx==0.25.2
python-dotenv==1.0.0